Split up massifreskin config for ease of maintenance.
